change test organization
